Move producer unit test to test/test_producer.py
